requirements.txt
pandas
openpyxl
pyyaml
seaborn